Move to CMake and remove libreadline...
CMake works more gracefully with more cross compilation frameworks,
letting us make this with all sorts of goofball stuff.  Autotools
doesn't let us do this, so OUT it goes...  (Not that I'd have
WILLINGLY kept the busted-*ss build system there, esp. for such
a small project as this and in the large they were **NOT** using
any of it's features in the large.  Sigh...what were people thinking?

While the licensing is compatible with libreadline, it's a pain for
embedded systems unless it's used in the target.  Moved to linenoise.
While this is a deviation from the debian upstreams, this makes it
static link capable (next commit I suspect) and able to drop onto a
variety of systems that would otherwise be problematic because they
wouldn't be using libreadline for license safety (In all honesty...a
good thing).  Linenoise is fixed in updates, etc. and is Two-clause
BSD licensed an is smaller, lighter and works MOSTLY like readline.
(Which means it can be just dropped in the project and GO.)